pyproject.toml
[tool.poetry]
name = "urlshortener"
version = "0.1.0"
description = ""
authors = ["Fernando López <fernando.e.lopez@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "^5.0.3"
django-rest-framework = "^0.1.0"
python-dotenv = "^1.0.1"
beautifulsoup4 = "^4.12.3"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
black = "^24.2.0"
isort = "^5.13.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
